+++ apps/remnant2toolkit/app/(actions)/builds/create-build.ts
@@ -64,6 +64,9 @@ export async function createBuild(data: string): Promise<BuildActionResponse> {
   const descriptionBadWordCheck = badWordFilter.isProfane(
     buildState.description ?? '',
   )
+  const referenceLinkBadWordCheck = badWordFilter.isProfane(
+    buildState.buildLink ?? '',
+  )
 
   if (nameBadWordCheck.isProfane) {
     buildState.isPublic = false
@@ -136,6 +139,45 @@ export async function createBuild(data: string): Promise<BuildActionResponse> {
     return {
       errors: [
         `Build description contains profanity: ${descriptionBadWordCheck.badWords.join(
+          ', ',
+        )}`,
+      ],
+    }
+  }
+
+  if (referenceLinkBadWordCheck.isProfane) {
+    buildState.isPublic = false
+
+    // Send webhook to #action-log
+    await sendWebhook({
+      webhook: 'auditLog',
+      params: {
+        embeds: [
+          {
+            title: `Bad Word Filter Tripped`,
+            color: 0xff0000,
+            fields: [
+              {
+                name: 'Action',
+                value: 'Create Build, Reference Link',
+              },
+              {
+                name: 'User',
+                value: session.user.displayName,
+              },
+              {
+                name: 'Bad Words',
+                value: referenceLinkBadWordCheck.badWords.join(', '),
+              },
+            ],
+          },
+        ],
+      },
+    })
+
+    return {
+      errors: [
+        `Reference link contains profanity: ${referenceLinkBadWordCheck.badWords.join(
           ', ',
         )}`,
       ],