+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-06-24
+
+### Added
+
+- Added bad word filter check to the build link field.
+
 ## 2024-06-23
 
 ### Fixed